Move function select  to the base module
